requirements.txt
certifi==2020.6.20
charset-normalizer==2.0.12
idna==3.3
oauthlib==3.2.0
py4j==0.10.9.3
pyspark==3.2.1
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
tweepy==4.8.0
urllib3==1.26.9
